Compile batch_fn
This will help us ensure we don't have shapes changing
throughout epochs. The current test suite fails exactly
because of that.

Since we are now compiling, instead of jitting, it may
be that the name `jit_compile?` is no longer relevant.
We should discuss if we want to keep the name altogether.